+++ src/main/resources/db/migrations/V8__create_shopping_list_table.sql
@@ -1,6 +1,6 @@
 create table shopping_list (
     id SERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
-    latitude FLOAT NOT NULL,
-    longitude FLOAT NOT NULL
+    latitude DOUBLE NOT NULL,
+    longitude DOUBLE NOT NULL
 );
